Challenge/Handons1/tasker/tests/test_task_service.py
import unittest
from app.services.task_service import TaskService

class TestTaskService(unittest.TestCase):
    def setUp(self):
        # Set up any necessary test data or dependencies
        pass

    def tearDown(self):
        # Clean up any resources used for testing
        pass

    def test_create_task(self):
        # Test creating a new task
        task_data = {
            "name": "Task 1",
            "description": "This is task 1",
            "priority": "High",
            "duedate": "2022-01-01",
            "user_id": 1
        }
        task_service = TaskService()
        created_task = task_service.create_task(task_data)
        self.assertIsNotNone(created_task)
        self.assertEqual(created_task["name"], "Task 1")
        # Add more assertions as needed

    def test_get_task(self):
        # Test retrieving a task by ID
        task_id = 1
        task_service = TaskService()
        task = task_service.get_task(task_id)
        self.assertIsNotNone(task)
        self.assertEqual(task["id"], task_id)
        # Add more assertions as needed

    def test_update_task(self):
        # Test updating an existing task
        task_id = 1
        updated_task_data = {
            "name": "Updated Task",
            "description": "This is an updated task",
            "priority": "Low",
            "duedate": "2022-02-01",
            "user_id": 1
        }
        task_service = TaskService()
        updated_task = task_service.update_task(task_id, updated_task_data)
        self.assertIsNotNone(updated_task)
        self.assertEqual(updated_task["name"], "Updated Task")
        # Add more assertions as needed

    def test_delete_task(self):
        # Test deleting a task
        task_id = 1
        task_service = TaskService()
        deleted_task = task_service.delete_task(task_id)
        self.assertIsNone(deleted_task)
        # Add more assertions as needed

if __name__ == "__main__":
    unittest.main()